Close to 0.2
