+++ book/src/SUMMARY.md
@@ -1,3 +1,6 @@
 # Summary
 
 - [Getting Started](./getting_started.md)
+- [Pathfinding](./pathfinding.md)
+- [Configuration](./configuration.md)
+- [Debugging](./debugging.md)
