+++ examples/collision_testing.rs
@@ -1,5 +1,7 @@
 use bevy::{
     dev_tools::fps_overlay::{FpsOverlayConfig, FpsOverlayPlugin},
+    log,
+    platform::collections::HashMap,
     prelude::*,
     text::FontSmoothing,
 };
@@ -34,11 +36,11 @@ fn main() {
         .add_plugins((TilemapPlugin, TiledMapPlugin::default()))
         // bevy_northstar plugins
         .add_plugins((
-            NorthstarPlugin::<OrdinalNeighborhood>::default(),
-            NorthstarDebugPlugin::<OrdinalNeighborhood>::default(),
+            NorthstarPlugin::<CardinalNeighborhood>::default(),
+            NorthstarDebugPlugin::<CardinalNeighborhood>::default(),
         ))
         // Add the SharedPlugin for unrelated pathfinding systems shared by the examples
-        .add_plugins(shared::SharedPlugin::<OrdinalNeighborhood>::default())
+        .add_plugins(shared::SharedPlugin::<CardinalNeighborhood>::default())
         // Observe the LayerCreated event to build the grid from the Tiled layer
         .add_observer(layer_created)
         // Startup and State Systems
@@ -60,7 +62,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
     // Get our anchor positioning calculated
     let anchor = TilemapAnchor::Center;
 
-    let tilemap_size = TilemapSize { x: 128, y: 128 };
+    let tilemap_size = TilemapSize { x: 16, y: 16 };
     let tilemap_gridsize = TilemapGridSize { x: 8.0, y: 8.0 };
     let tilemap_tilesize = TilemapTileSize { x: 8.0, y: 8.0 };
 
@@ -78,28 +80,33 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
     );
 
     // Spawn a 2D camera and set the position based on the centered anchor offset
-    commands.spawn(Camera2d).insert(Transform {
-        translation: camera_offset,
-        ..Default::default()
-    });
+    commands.spawn(Camera2d).insert((
+        Transform {
+            translation: camera_offset,
+            ..Default::default()
+        },
+        Projection::Orthographic(OrthographicProjection {
+            scale: 0.01,
+            ..OrthographicProjection::default_2d()
+        }),
+    ));
 
     // Load the map ...
-    let map_handle: Handle<TiledMap> = asset_server.load("demo_128.tmx");
+    let map_handle: Handle<TiledMap> = asset_server.load("demo_16.tmx");
 
     let mut map_entity = commands.spawn((TiledMapHandle(map_handle), anchor));
 
-    // Insert the grid as a child of the map entity. This won't currently affect anything, but in the future
-    // we may want to have the grid as a child of the map entity so that multiple grids can be supported.
+    // You can eventually add some extra settings to your map
     map_entity.insert((
         TilemapRenderSettings {
             render_chunk_size: UVec2::new(32, 32),
             ..Default::default()
         },
-        Grid::<OrdinalNeighborhood>::new(&GridSettings {
-            width: 128,
-            height: 128,
+        Grid::<CardinalNeighborhood>::new(&GridSettings {
+            width: 16,
+            height: 16,
             depth: 1,
-            chunk_size: 16,
+            chunk_size: 8,
             chunk_depth: 1,
             chunk_ordinal: false,
             default_cost: 1,
@@ -109,16 +116,14 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
         }),
     ));
 
-    // Add the debug map as a child of the map entity
-    // Set the translation to offset the the debug gizmos.
     map_entity.with_child((
         DebugMap {
             tile_width: 8,
             tile_height: 8,
             map_type: DebugMapType::Square,
-            draw_chunks: true,
+            draw_chunks: false,
             draw_points: false,
-            draw_entrances: true,
+            draw_entrances: false,
             draw_cached_paths: false,
         },
         Transform::from_translation(offset.extend(0.0)),
@@ -128,7 +133,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
 fn layer_created(
     trigger: Trigger<TiledLayerCreated>,
     map_asset: Res<Assets<TiledMap>>,
-    grid: Single<&mut Grid<OrdinalNeighborhood>>,
+    grid: Single<&mut Grid<CardinalNeighborhood>>,
     mut state: ResMut<NextState<shared::State>>,
 ) {
     let mut grid = grid.into_inner();
@@ -170,11 +175,10 @@ fn layer_created(
 
 fn spawn_minions(
     mut commands: Commands,
-    grid: Query<&Grid<OrdinalNeighborhood>>,
+    grid: Query<&Grid<CardinalNeighborhood>>,
     layer_entity: Query<Entity, With<TiledMapTileLayer>>,
     asset_server: Res<AssetServer>,
     mut walkable: ResMut<shared::Walkable>,
-    config: Res<shared::Config>,
 ) {
     let grid = if let Ok(grid) = grid.single() {
         grid
@@ -196,10 +200,10 @@ fn spawn_minions(
     }
 
     let mut count = 0;
+    let mut rng = rand::rng();
 
-    while count < 128 {
-        let position = walkable.tiles.choose(&mut rand::rng()).unwrap();
-        let goal = walkable.tiles.choose(&mut rand::rng()).unwrap();
+    while count < 6 {
+        let position = walkable.tiles.choose(&mut rng).unwrap();
 
         let transform = Vec3::new(position.x + 4.0, position.y + 4.0, 4.0);
 
@@ -231,10 +235,6 @@ fn spawn_minions(
                 (position.y / 8.0) as u32,
                 0,
             )))
-            .insert(Pathfind {
-                goal: UVec3::new((goal.x / 8.0) as u32, (goal.y / 8.0) as u32, 0),
-                use_astar: config.use_astar,
-            })
             .insert(ChildOf(layer_entity));
 
         count += 1;
@@ -255,9 +255,23 @@ fn move_pathfinders(
     let (map_size, tile_size, grid_size, anchor) = tilemap.into_inner();
 
     for _ in tick_reader.read() {
+        let mut debug_next = HashMap::new();
+
         let offset = anchor.as_offset(map_size, grid_size, tile_size, &TilemapType::Square);
 
         for (entity, mut position, next) in query.iter_mut() {
+            if debug_next.contains_key(&next.0) {
+                log::error!(
+                    "Entity {:?} has the same next position as another entity: {:?} {:?}",
+                    entity,
+                    next,
+                    debug_next.get(&next.0)
+                );
+                continue;
+            }
+
+            debug_next.insert(next.0, entity);
+
             position.0 = next.0;
 
             let translation = Vec3::new(
@@ -278,14 +292,13 @@ fn set_new_goal(
     mut commands: Commands,
     mut minions: Query<Entity, (Without<Path>, Without<Pathfind>)>,
     walkable: Res<shared::Walkable>,
-    config: Res<shared::Config>,
 ) {
     for entity in minions.iter_mut() {
         let new_goal = walkable.tiles.choose(&mut rand::rng()).unwrap();
 
         commands.entity(entity).insert(Pathfind {
             goal: UVec3::new((new_goal.x / 8.0) as u32, (new_goal.y / 8.0) as u32, 0),
-            use_astar: config.use_astar,
+            use_astar: false,
         });
     }
 }
@@ -293,6 +306,7 @@ fn set_new_goal(
 fn handle_reroute_failed(
     mut commands: Commands,
     mut query: Query<(Entity, &Pathfind, &RerouteFailed)>,
+    config: Res<shared::Config>,
     mut tick_reader: EventReader<shared::Tick>,
 ) {
     for _ in tick_reader.read() {
@@ -300,7 +314,7 @@ fn handle_reroute_failed(
             commands.entity(entity).remove::<RerouteFailed>();
             commands.entity(entity).insert(Pathfind {
                 goal: pathfind.goal,
-                use_astar: true,
+                use_astar: config.use_astar,
             });
         }
     }
